+++ ergo-nipopow/src/nipopow_proof.rs
@@ -9,12 +9,9 @@ use sigma_ser::{
 };
 use std::convert::TryFrom;
 
-use crate::{
-    autolykos_pow_scheme::{self},
-    nipopow_algos::NipopowAlgos,
-};
+use crate::{autolykos_pow_scheme, nipopow_algos::NipopowAlgos};
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
 /// A structure representing NiPoPow proof as a persistent modifier.
 pub struct NipopowProof {
     /// Algos
@@ -283,5 +280,76 @@ impl ScorexSerializable for PoPowHeader {
             interlinks: interlinks?,
             interlinks_proof: interlinks_proof?,
         })
+    }
+}
+
+#[cfg(feature = "arbitrary")]
+#[allow(clippy::unwrap_used)]
+mod arbitrary {
+    use crate::autolykos_pow_scheme::AutolykosPowScheme;
+
+    use super::*;
+    use ergo_chain_types::Digest32;
+    use proptest::prelude::*;
+    use proptest::{arbitrary::Arbitrary, collection::vec};
+
+    impl Arbitrary for PoPowHeader {
+        type Parameters = ();
+        type Strategy = BoxedStrategy<PoPowHeader>;
+
+        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
+            (any::<Box<Header>>(), vec(any::<Digest32>(), 1..10))
+                .prop_map(|(header, digests)| PoPowHeader {
+                    header: *header,
+                    interlinks: digests.into_iter().map(BlockId).collect(),
+                })
+                .boxed()
+        }
+    }
+
+    impl Arbitrary for NipopowProof {
+        type Parameters = ();
+        type Strategy = BoxedStrategy<NipopowProof>;
+
+        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
+            (
+                any::<u32>(),
+                any::<u32>(),
+                vec(any::<PoPowHeader>(), 1..10),
+                any::<PoPowHeader>(),
+                vec(any::<Header>(), 1..10),
+            )
+                .prop_map(|(m, k, prefix, suffix_head, suffix_tail)| NipopowProof {
+                    popow_algos: NipopowAlgos {
+                        pow_scheme: AutolykosPowScheme::default(),
+                    },
+                    m,
+                    k,
+                    prefix,
+                    suffix_head,
+                    suffix_tail,
+                })
+                .boxed()
+        }
+    }
+}
+
+#[cfg(test)]
+#[cfg(feature = "arbitrary")]
+#[allow(clippy::unwrap_used, clippy::panic)]
+pub mod tests {
+    use super::*;
+    use proptest::prelude::*;
+    use sigma_ser::scorex_serialize_roundtrip;
+    proptest! {
+
+        #![proptest_config(ProptestConfig::with_cases(64))]
+
+        #[test]
+        fn nipopowproof_roundtrip(v in any::<NipopowProof>()) {
+            prop_assert_eq![scorex_serialize_roundtrip(&v), v];
+        }
+
+
     }
 }