+++ ergo-rest/src/api/node.rs
@@ -92,6 +92,6 @@ mod tests {
             timeout: Some(Duration::from_secs(5)),
         };
         let res = runtime_inner.block_on(async { get_info(node_conf).await.unwrap() });
-        assert_eq!(res.name, "ergo-mainnet-4.0.16.1");
+        assert_ne!(res.name, "");
     }
 }